improve calculation accuracy + fix truncation error
Allow using floating point arguments to avoid truncation.
Test: Add (4.9, 4.9) should return 10.

+++ calculator.h
@@ -1,6 +1,6 @@
 class Calculator
 {
     public:
-        int Add (int, int);
-        int Sub (int, int);
+        int Add (double, double);
+        int Sub (double, double);
 };
